feat(user): 공통 상세 코드 Repository 생성 및 Data JPA 적용
Adds S09P31C108-134

+++ server/user-service/src/main/java/com/everyschool/userservice/domain/codegroup/repository/CodeGroupRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CodeGroupRepository extends JpaRepository<CodeGroup, Long> {
+public interface CodeGroupRepository extends JpaRepository<CodeGroup, Integer> {
 }
